Split local import resolution helpers
