+++ src/module_system/import_resolution/stdlib_paths.rs
@@ -1,3 +1,5 @@
+use std::path::PathBuf;
+
 use super::*;
 
 impl ModuleSystem {
